fsdp: consolidate the routing-replay hooks and trim commentary
