+++ miles/backends/experimental/fsdp_utils/adaptations/specs/qwen3_5.py
@@ -2,6 +2,7 @@
 cu_seqlens to fla chunk/recurrent_gated_delta_rule and seq_idx to causal_conv1d_fn per packed document.
 Patches the DecoderLayer/GatedDeltaNet class forwards; kernel logic lives in ``models/qwen3_5.py``."""
 
+from ...models.routing_replay import install_qwen3_router_replay
 from ..packing.registry import PackingPatch, register_packing_patch
 from ..routing_replay import RoutingReplayAdapter, register_routing_replay_adapter
 
@@ -26,17 +27,11 @@ register_packing_patch(PackingPatch("gated_deltanet_packing", _applies, "config"
 
 
 def _is_qwen3_5_moe(hf_config) -> bool:
-    """Qwen3.5 MoE. The text-only checkpoint reports model_type qwen3_5_moe_text."""
+    """Qwen3.5 MoE; the text-only checkpoint reports model_type qwen3_5_moe_text."""
     if hf_config is None:
         return False
     model_type = str(getattr(hf_config, "model_type", "") or "")
     return "qwen3_5" in model_type and "moe" in model_type
-
-
-def _install_routing_replay(module) -> None:
-    from ...models.routing_replay_qwen3 import install_qwen3_topk_router_replay
-
-    install_qwen3_topk_router_replay(module)
 
 
 register_routing_replay_adapter(
@@ -44,6 +39,6 @@ register_routing_replay_adapter(
         name="qwen3_5_moe",
         applies_to=_is_qwen3_5_moe,
         module_cls_name="Qwen3_5MoeTopKRouter",
-        install=_install_routing_replay,
+        install=install_qwen3_router_replay,
     )
 )
