+++ tests/fast/backends/test_fsdp_r3_orchestration.py
@@ -23,7 +23,6 @@ def _reset_manager():
 
 
 def test_enabled_follows_use_routing_replay():
-    # --use-rollout-routing-replay sets use_routing_replay during arg validation.
     assert resolve_routing_replay_enabled(Namespace(use_routing_replay=True)) is True
     assert resolve_routing_replay_enabled(Namespace(use_routing_replay=False)) is False
 
@@ -48,8 +47,6 @@ def test_stage_context_restores_on_exception():
 
 
 def test_stage_context_nests():
-    # The training step sets replay_forward inside a replay_backward region; leaving the
-    # forward must put the region back so backward recompute reads the backward cursor.
     routing_replay_manager.stage = "fallthrough"
     with routing_replay_stage("replay_backward"):
         assert routing_replay_manager.stage == "replay_backward"
@@ -60,9 +57,6 @@ def test_stage_context_nests():
 
 
 def test_ref_model_creation_does_not_install_routing_replay():
-    # The ref pass runs as fallthrough. If _create_ref_model ever installed R3, the ref
-    # model's Replay objects would land in the same manager.replays list, doubling its
-    # length and invalidating every stream_idx during fill_replay_data.
     import inspect
 
     from miles.backends.experimental.fsdp_utils.actor import FSDPTrainRayActor
